Fixed timing violations
All timing violations had to do with crossing clock domains between
the I2C slave and the various entities that the registers controlled.
Upon examination, we determined that most of these would not be an
issue because the entities were coded well (had syncing), but the
MonADC controller (COTS_AD7274.vhd) did not, and so we added the
signals start_s1 and start_s2. After that was fixed, we were able
to put all clocks in a clock group.